% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.R
\docType{data}
\name{births}
\alias{births}
\title{Births in England}
\format{A data frame with 102 rows and 3 variables.
\describe{
\item{year}{Year of birth}
\item{births}{The number of births per year}
\item{data}{Whether the data has been observed or projected}
}}
\usage{
births
}
\description{
A dataset containing the observed number of births for 2000-2015 from the Office of National
Statistics (ONS), from 2015-2101 the number of births have been projected again by the ONS.
}
\keyword{datasets}
