% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england_demographics.R
\docType{data}
\name{england_demographics}
\alias{england_demographics}
\title{Population data stratified by age group and UK birth status; Extracted, Cleaned, and Aggregated from the LFS}
\format{A data frame with 386 rows and 6 variables.
\describe{
\item{year}{Year}
\item{age_group}{Age group to be used in modelling}
\item{population}{Population as estimated from the LFS}
}}
\usage{
england_demographics
}
\description{
A dataset containing population data stratified by birth status, by year from 2000 to 2016, estimmated using the Labour Force Survey.
See \href{https://www.samabbott.co.uk/tbinenglanddataclean/}{tbinenglanddataclean}.
for information regarding the raw data sources used, extraction, and cleaning.
}
\keyword{datasets}
