% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everything_from_model.R
\name{everything_from_model}
\alias{everything_from_model}
\title{Adapt LibBi model with all Variables in Output}
\usage{
everything_from_model(model)
}
\arguments{
\item{model}{A LibBi model, loaded using \code{bi_model}.}
}
\value{
A LibBi model.
}
\description{
Some \code{rbi.helper} functions require all variables to have outputs. This function modifies a LibBi model
so that this is the case.
}
\examples{
library(rbi)

model <- system.file(package="rbi", "SIR.bi") # get full file name from package
model <- bi_model(model) # load model
model <- replace_all(model, "noise n_recovery", "noise n_recovery\\\\(has_output = 0\\\\)")

model

everything_from_model(model)
}
