% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit Model}
\usage{
fit_model(model = "BaseLineModel", previous_model_path = NULL,
  gen_data = TRUE, run_time = 73, time_scale = "year",
  plot_obs = TRUE, sample_priors = TRUE, prior_samples = 1000,
  nparticles = NULL, adapt_particles = FALSE,
  adapt_part_samples = 1000, adapt_part_it = 10, min_particles = 4,
  max_particles = 16, proposal_param_block = NULL,
  proposal_initial_block = NULL, adapt_proposal = FALSE,
  adapt_prop_samples = 1000, adapt_prop_it = 10, adapt = "both",
  adapt_scale = 2, min_acc = 0.2, max_acc = 0.4, fit = FALSE,
  posterior_samples = 10000, thin = 10, burn_prop = 0,
  nthreads = parallel::detectCores(), verbose = TRUE,
  libbi_verbose = FALSE, fitting_verbose = TRUE, browse = FALSE,
  const_pop = FALSE, no_age = FALSE, no_disease = FALSE,
  save_output = FALSE, dir_path = NULL, dir_name = NULL,
  seed = 1234)
}
\arguments{
\item{model}{A character string containing the model name. Alternatively a model loaded via \code{rbi::bi_model} may be passed.}

\item{previous_model_path}{A character string giving the path to a previous LiBBi model (saved as an .rds). This will replace the default model and
override initial settings.}

\item{gen_data}{Logical, defaults to /code{TRUE}. Should data be synthesised using the model and priors.}

\item{run_time}{Numeric, the number of years to run the model fitting and simulation for, defaults to 74 (i.e from 1931 until 2005).}

\item{time_scale}{Character, defaults to \code{"year"}. A monthly timescale can also be set with \code{"month" }.}

\item{plot_obs}{Logical, defaults to \code{TRUE}. Should input data be plotted}

\item{sample_priors}{Logical, defaults to \code{FALSE}. Should the model priors be sampled.}

\item{prior_samples}{Numeric, the number of samples to take from the priors. Defaults to 1000.}

\item{nparticles}{Numeric, the initial number of particles to use in the particle filters. Defaults to \code{NULL}, in which case the number of outputs
is used as the initial particle number.}

\item{adapt_particles}{Logical, defaults to \code{FALSE}. Should the number of particles be adapted.}

\item{adapt_part_samples}{Numeric, the number of samples to take from the priors when adapting the number of particles. Defaults to 1000.}

\item{adapt_part_it}{Numeric, the number of iterations to use for adapting the number of particles. Defaults to 10.}

\item{min_particles}{Numeric, defaults to 4. The starting number of particles to use for adpation.}

\item{max_particles}{Numeric, defaults to 16. The maximum number of particles to use for apation.}

\item{proposal_param_block}{A character string containing a LiBBi proposal parameter block. Defaults to \code{NULL} in
which case the LiBBi defaults will be used.}

\item{proposal_initial_block}{A character string containing a LiBBi proposal initial block. Defaults to \code{NULL} in
which case the LiBBi defaults will be used.}

\item{adapt_proposal}{Logical, defaults to \code{TRUE}. Should the proposal be adjusted to improve the acceptance rate.}

\item{adapt}{Character string, defaults to "both". The type of adaption to use for the proposal see \code{rbi.helpers::adapt_proposal} for details.}

\item{adapt_scale}{Numeric, defaults to 2. The scale to use to adapt the proposal.}

\item{min_acc}{Numeric, defaults to 0.05. The minimum target acceptance rate.}

\item{max_acc}{Numeric, defaults to 0.4. The maximum target acceptance rate.}

\item{fit}{Logical, defaults to \code{TRUE}. Should the model be fitted with 1000 samples extracted.}

\item{posterior_samples}{Numeric, the number of samples to take from the posterior estimated using pmcmc (requires \code{fit = TRUE}). Defaults to 1000.}

\item{thin}{Numeric, the thinning interval to use between posterior samples. May reduce the correlation between samples and reduces memory.}

\item{burn_prop}{Numeric (between 0 and 1). The proportion of the chain to discard as burn-in.}

\item{nthreads}{Numeric, defaults to 4. The number of parallel jobs to run. The most efficient option is likely to be to match the
number of cores available.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should progress messages and output be printed.}

\item{libbi_verbose}{Logical, defaults to \code{FALSE}. Should \code{libbi} produce verbose progress and warnings messages.}

\item{fitting_verbose}{Logical, defaults to \code{FALSE}. Should \code{libbi} produce verbose progress and warnings messages whilst fitting.}

\item{browse}{Logical, defaults to \code{FALSE}. Should the function be run in debug mode using \code{browser}.}

\item{const_pop}{Logical, defaults to \code{FALSE}. Should a constant population be maintained using births equals deaths.}

\item{no_age}{Logical, defaults to \code{FALSE}. Should ageing be disabled.}

\item{no_disease}{Logical, defaults to \code{FALSE}. Should disease be removed from the model}

\item{save_output}{Logical, defaults to \code{FALSE}. Should the model results be saved as a test dataset.}

\item{seed}{Numeric, the seed to use for random number generation.}

\item{adapt_part_samples}{Numeric, the number of samples to take from the priors when adapting the proposal distribution. Defaults to 1000.}

\item{adapt_part_it}{Numeric, defaults to 10. The number of iterations to use for adapting the proposal.}
}
\value{
A LibBi model object based on the inputed test model.
}
\description{
A function to fit the models in development through the full fitting pipeline based on synthetic data or observed data.
}
\examples{

## Function code:
fit_model
}
