% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\docType{data}
\name{incidence}
\alias{incidence}
\title{TB Incidence Data; Extracted, Cleaned, and Aggregated from the ETS}
\format{A data frame with - rows and 6 variables.
\describe{
\item{year}{Year of notification}
\item{age_group}{Age group to be used in modelling, aggregated from age at notification}
\item{ukborn}{UK birth status}
\item{pulmextrapulm}{Pulmonary/Extra-Pulmonary status of notified cases}
\item{type}{Data subset into data for fitting or observation.}
\item{incidence}{Incidence in the given timeframe}
\item{policy_change}{Was the notification observed before or after the change in vaccination policy.}
}}
\usage{
incidence
}
\description{
A dataset containing TB incidence data by year from 2000 to 2015.
}
\keyword{datasets}
