% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{model_plots}
\alias{model_plots}
\title{Model Summary Plots}
\usage{
model_plots(model, uncertainty = TRUE, prior = NULL)
}
\arguments{
\item{model}{A LibBi model object}

\item{uncertainty}{Logical, defaults to \code{TRUE}.}

\item{prior}{A libBi model object. Containing priors to be compared to. Defaults to \code{NULL}}
}
\value{
A list of model summary plots
}
\description{
Model Summary Plots
}
\examples{


## Code 
model_plots
}
\seealso{
plot_param plot_state
}
