% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTransVsDirectBCG.R
\docType{package}
\name{ModelTBBCGEngland-package}
\alias{ModelTBBCGEngland}
\alias{ModelTBBCGEngland-package}
\title{ModelTBBCGEngland: Fit a TB model, including BCG vaccination, to available routine surveillance data in England}
\description{
Fit a dynamic model of TB, including BCG vaccination, to available routine surveillance data in England.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.samabbott.co.uk/ModelTBBCGEngland}
  \item \url{https://github.com/seabbs/ModelTBBCGEngland}
  \item Report bugs at \url{https://github.com/seabbs/ModelTBBCGEngland/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk}

}
\keyword{internal}
