% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rates.R
\docType{data}
\name{mortality_rates}
\alias{mortality_rates}
\title{Yearly Age Specific Mortality Rates}
\format{A data frame with - rows and 4 variables.
\describe{
\item{year}{Calendar year}
\item{age_group}{5 year age groups (0-4, 5-9, ...., 85-89, 90+)}
\item{mortality_rate}{Estimated age specific mortality rate per year}
\item{exp_life_span}{Estimated expected lifespan if an individual continued to experiance
the same level of age specific mortality}
}}
\usage{
mortality_rates
}
\description{
A dataset containing estimated age specific mortality rates from 2000 until 2014. These
are estimated by the ONS and are 3 year rolling estimates. See \href{www.samabbott.co.uk/tbinenglanddataclean}{tbinenglanddataclean}
for details.
}
\keyword{datasets}
