% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param.R
\name{plot_param}
\alias{plot_param}
\title{Plot a Subset of Parameters From a LiBBi Model}
\usage{
plot_param(libbi_model = NULL, param = NULL, prior_params = NULL,
  log_scale = FALSE, sqrt_scale = FALSE, plot_type = "dist",
  plot_data = TRUE, burn_in = 0, scales = "fixed")
}
\arguments{
\item{libbi_model}{LiBBi model object}

\item{param}{A character vector containing the complete names of the variables to plot.}

\item{prior_params}{A prior parameter set as generated by \code{bi_read}.}

\item{log_scale}{Logical, defaults to \code{TRUE}. Should the x axis be plotted on a log scale?}

\item{sqrt_scale}{Logical, defaults to \code{TRUE}. Should the x axis be plotted on a square root scale?}

\item{plot_type}{A character string indicating the type of plot to show. \code{"dist"} plots a density and rug plot, whilst \code{"trace"}
plots a trace plot.}

\item{plot_data}{Logical, defaults to \code{TRUE}. Should the summarised data be plotted.}

\item{burn_in}{Numeric, defaults to 0. The number of samples to remove as burn in.}

\item{scales}{Character string, defaulting to "fixed". What scales to use for graph faceting.}
}
\value{
A plot of the specified parameters.
}
\description{
Plot either the distribution or trace of a subset of parameters split out by their dimension
}
\examples{

##Show function code
plot_param
}
