% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensitivity.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot output sensitivty to parameters}
\usage{
plot_sensitivity(sensitivity = NULL)
}
\arguments{
\item{sensitivity}{A dataframe of sensitivity data as produced by \code{test_sensitivity}.}
}
\value{
A plot summarising model sensitivity to parameter changes.
}
\description{
Plot output sensitivty to parameters
}
\examples{

## Code
plot_sensitivity
}
