% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state.R
\name{plot_state}
\alias{plot_state}
\title{Plot a Subset of States From a LiBBi Model}
\usage{
plot_state(libbi_model = NULL, states = NULL, summarise = FALSE,
  summarise_by = NULL, strat_var = "state")
}
\arguments{
\item{libbi_model}{LiBBi model object}

\item{states}{A character vector containing the complete names of the variables to plot.}
}
\value{
A plot of the specified states.
}
\description{
Plot a subset of the states from a LiBBi model and summarising multiple runs.
Multi dimension states can be plotted using stratification and facetting.
}
\examples{


}
