% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_libbi.R
\name{read_libbi}
\alias{read_libbi}
\title{Read a LiBbi Model}
\usage{
read_libbi(folder, ...)
}
\arguments{
\item{folder}{Name of the folder containing the Libbi output as formated by \code{save_libbi}.}

\item{...}{any extra options to pass to \code{\link{read_libbi}} when creating the new object}
}
\value{
a \code{\link{libbi}} object#'
}
\description{
Read results of a \code{LibBi} run from a folder. This completely reconstructs the saved \code{LibBi} object
This reads all options, files and outputs of a \code{LibBi} run from a specified folder
}
\examples{

## Function code
ModelTBBCGEngland::read_libbi
}
