% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_code_with_form.R
\name{replace_code_with_form}
\alias{replace_code_with_form}
\title{Replace Coded Parameter names with latex.}
\usage{
replace_code_with_form(param)
}
\arguments{
\item{param}{A vector of coded parameters.}
}
\value{
A vector of parameters formated with latex
}
\description{
Replace Coded Parameter names with latex.
}
\examples{


}
