% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_libbi.R
\name{save_libbi}
\alias{save_libbi}
\title{Save a LiBbi Model}
\usage{
save_libbi(x, folder, supplement)
}
\arguments{
\item{x}{a \code{\link{libbi}} object}

\item{folder}{A character string indicating the folder name under which to save the model.}

\item{supplement}{any supplementary data to save}
}
\description{
Write results of a \code{LibBi} run to a folder as a series of RDS files.
This saves all options, files and outputs of a \code{LibBi} run to a specified folder.
}
\examples{

## Function code 
save_libbi
}
