% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model_obs.R
\name{setup_model_obs}
\alias{setup_model_obs}
\title{Setup Observed Model Data}
\usage{
setup_model_obs(years_of_age = NULL)
}
\arguments{
\item{years_of_age}{Numeric, the years of age distributed cases to fit to.
If not specified then no years are used.}
}
\value{
A named list of observed data required by the model.
}
\description{
Performs data preprocessing required to prepare the observed data
for the model.
}
\examples{

## Code
setup_model_obs

## Output
setup_model_obs()

}
