% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_model}
\alias{test_model}
\title{Test Model}
\usage{
test_model(model = "BaseLineModel", nparticles = 1,
  adapt_particles = FALSE, adapt_proposal = FALSE, min_acc = 0.05,
  max_acc = 0.4, test_fit = TRUE, save_output = FALSE, verbose = TRUE,
  browse = FALSE)
}
\arguments{
\item{model}{A character string containing the model name. Alternatively a model loaded via \code{rbi::bi_model} may be passed.}

\item{nparticles}{Numeric, the initial number of particles to use in the particle filters.}

\item{adapt_particles}{Logical, defaults to \code{FALSE}. Should the number of particles be adapted.}

\item{adapt_proposal}{Logical, defaults to \code{TRUE}. Should the proposal be adjusted to improve the acceptance rate.}

\item{min_acc}{Numeric, defaults to 0.05. The minimum target acceptance rate.}

\item{max_acc}{Numeric, defaults to 0.4. The maximum target acceptance rate.}

\item{test_fit}{Logical, defaults to \code{TRUE}. Should the model be fitted with 1000 samples extracted.}

\item{save_output}{Logical, defaults to \code{FALSE}. Should the model results be saved as a test dataset.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should progress messages and output be printed.}

\item{browse}{Logical, defaults to \code{FALSE}. Should the function be run in debug mode using \code{browser}.}
}
\value{
A LibBi model object based on the inputed test model.
}
\description{
A function to test the models in development through the full fitting pipeline based on synthetic data.
}
\examples{


}
