% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sensitivity.R
\name{test_sensitivity}
\alias{test_sensitivity}
\title{Test the sensitivity of a Libbi model using PRCC}
\usage{
test_sensitivity(model = NULL, obs = NULL, target_time = NULL)
}
\arguments{
\item{model}{A LiBBi model object}

\item{obs}{A character string containing the observation to use to calculate the PRCC.}

\item{target_time}{Numeric, the time at which to estimate the model sensitivity. If not specified
then this will default to the last fitted point that the model has produced output for.}
}
\value{
A data frame containing the names of the parameters in the model, the correlation with the outcome
and the p value of this correlation
}
\description{
Test the sensitivity of a Libbi model using PRCC
}
\examples{

test_sensitivity
}
