% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param.R
\name{plot_param}
\alias{plot_param}
\title{Plot a Subset of Parameters From a LiBBi Model}
\usage{
plot_param(libbi_model = NULL, parameter = NULL, log_scale = FALSE,
  sqrt_scale = FALSE, plot_type = "dist")
}
\arguments{
\item{libbi_model}{LiBBi model object}

\item{parameter}{A character vector containing the complete names of the variables to plot.}

\item{log_scale}{Logical, defaults to \code{TRUE}. Should the x axis be plotted on a log scale?}

\item{sqrt_scale}{Logical, defaults to \code{TRUE}. Should the x axis be plotted on a square root scale?}

\item{plot_type}{A character string indicating the type of plot to show. \code{"dist"} plots a density and rug plot, whilst \code{"trace"}
plots a trace plot.}
}
\value{
A plot of the specified parameters.
}
\description{
Plot either the distribution or trace of a subset of parameters split out by their dimension
}
\examples{


}
