% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state.R
\name{plot_state}
\alias{plot_state}
\title{Plot a Subset of States From a LiBBi Model}
\usage{
plot_state(libbi_model = NULL, model_paths = NULL, states = NULL,
  plot_obs = TRUE, obs = NULL, summarise = FALSE,
  summarise_by = NULL, strat_var = "state", burn_in = 0,
  start_time = 0, start_time_label = 1931, scenarios_start = 2005,
  scales = "free_y", plot_uncert = TRUE, plot_data = TRUE,
  show_mean = FALSE, use_comma_formatting = TRUE)
}
\arguments{
\item{libbi_model}{LiBBi model object (or a list of named libbi models).}

\item{model_paths}{A named character vector of paths to models to plot.}

\item{states}{A character vector containing the complete names of the variables to plot. Defaults to all states}

\item{plot_obs}{Logical, defaults to \code{TRUE}. Should the observed data be plotted.}

\item{obs}{List of dataframes containing observed data. Defaults to generating observed
data using \code{setup_model_obs}.}

\item{burn_in}{Numeric, indicating the burn in period (samples).}

\item{start_time}{Numeric, indicating the time to start plotting from.}

\item{start_time_label}{Numeric, the label to apply to the time variable. Defaults to 1931.}

\item{scenarios_start}{Numeric, start time for which to plot alternative scenarios. Defaults to 2005.}

\item{scales}{A character string indicating the axis scaling to use for facets. Defaults to
"free_y".}

\item{plot_uncert}{Logical, defaults to \code{TRUE}. Should simulation uncertainty be plotted.}

\item{plot_data}{Logical, defaults to \code{TRUE}. Should the summarised data be plotted.}

\item{show_mean}{Logical, defaults to \code{TRUE}. Should the mean value as well as the median be plotted.}

\item{use_comma_formatting}{Logicial, defaults to \code{TRUE}. Should comma formating be used on the y axis.}
}
\value{
A plot of the specified states.
}
\description{
Plot a subset of the states from a LiBBi model and summarising multiple runs.
Multi dimension states can be plotted using stratification and facetting.
}
\examples{

## Show function code
plot_state
}
