% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state.R
\name{plot_state}
\alias{plot_state}
\title{Plot a Subset of States From a LiBBi Model}
\usage{
plot_state(libbi_model = NULL, states = NULL, plot_obs = TRUE,
  obs = NULL, summarise = FALSE, summarise_by = NULL,
  strat_var = "state", burn_in = 0, scales = "free_y",
  plot_uncert = TRUE, plot_data = TRUE)
}
\arguments{
\item{libbi_model}{LiBBi model object}

\item{states}{A character vector containing the complete names of the variables to plot. Defaults to all states}

\item{plot_obs}{Logical, defaults to \code{TRUE}. Should the observed data be plotted.}

\item{obs}{List of dataframes containing observed data. Defaults to generating observed
data using \code{setup_model_obs}.}

\item{burn_in}{Numeric, indicating the burn in period.}

\item{scales}{A character string indicating the axis scaling to use for facets. Defaults to
"free_y".}

\item{plot_uncert}{Logical, defaults to \code{TRUE}. Should simulation uncertainty be plotted.}

\item{plot_data}{Logical, defaults to \code{TRUE}. Should the summarised data be plotted.}
}
\value{
A plot of the specified states.
}
\description{
Plot a subset of the states from a LiBBi model and summarising multiple runs.
Multi dimension states can be plotted using stratification and facetting.
}
\examples{

## Show function code
plot_state
}
