% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_model}
\alias{test_model}
\title{Test Model}
\usage{
test_model(model = "BaseLineModel", gen_data = TRUE, run_time = 74,
  time_scale = "year", plot_input_data = TRUE, sample_priors = TRUE,
  prior_samples = 1000, nparticles = NULL, adapt_particles = FALSE,
  adapt_proposal = FALSE, min_acc = 0.05, max_acc = 0.4, fit = FALSE,
  posterior_samples = 1000, save_output = FALSE, nthreads = 4,
  verbose = TRUE, libbi_verbose = FALSE, browse = FALSE,
  const_pop = FALSE, no_age = FALSE, no_disease = FALSE)
}
\arguments{
\item{model}{A character string containing the model name. Alternatively a model loaded via \code{rbi::bi_model} may be passed.}

\item{gen_data}{Logical, defaults to /code{TRUE}. Should data be synthesised using the model and priors.}

\item{run_time}{Numeric, the number of years to run the model fitting and simulation for, defaults to 74 (i.e from 1931 until 2005).}

\item{time_scale}{Character, defaults to \code{"year"}. A monthly timescale can also be set with \code{"month" }.}

\item{plot_input_data}{Logical, defaults to \code{TRUE}. Should input data be plotted}

\item{sample_priors}{Logical, defaults to \code{FALSE}. Should the model priors be sampled.}

\item{prior_samples}{Numeric, the number of samples to take from the priors. Defaults to 1000.}

\item{nparticles}{Numeric, the initial number of particles to use in the particle filters. Defaults to \code{NULL}, in which case the number of outputs
is used as the initial particle number.}

\item{adapt_particles}{Logical, defaults to \code{FALSE}. Should the number of particles be adapted.}

\item{adapt_proposal}{Logical, defaults to \code{TRUE}. Should the proposal be adjusted to improve the acceptance rate.}

\item{min_acc}{Numeric, defaults to 0.05. The minimum target acceptance rate.}

\item{max_acc}{Numeric, defaults to 0.4. The maximum target acceptance rate.}

\item{fit}{Logical, defaults to \code{TRUE}. Should the model be fitted with 1000 samples extracted.}

\item{posterior_samples}{Numeric, the number of samples to take from the posterior estimated using pmcmc (requires \code{fit = TRUE}). Defaults to 1000.}

\item{save_output}{Logical, defaults to \code{FALSE}. Should the model results be saved as a test dataset.}

\item{nthreads}{Numeric, defaults to 4. The number of parallel jobs to run. The most efficient option is likely to be to match the
number of cores available.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should progress messages and output be printed.}

\item{libbi_verbose}{Logical, defaults to \code{FALSE}. Should \code{libbi} produce verbose progress and warnings messages.}

\item{browse}{Logical, defaults to \code{FALSE}. Should the function be run in debug mode using \code{browser}.}

\item{const_pop}{Logical, defaults to \code{FALSE}. Should a constant population be maintained using births equals deaths.}

\item{no_age}{Logical, defaults to \code{FALSE}. Should ageing be disabled.}

\item{no_disease}{Logical, defaults to \code{FALSE}. Should disease be removed from the model}
}
\value{
A LibBi model object based on the inputed test model.
}
\description{
A function to test the models in development through the full fitting pipeline based on synthetic data.
}
\examples{


}
